// src/types/microshift.ts
export type Mircroshift = {
  stKeyName: string;        
  createdAt: string;         
  goal: string;              
  text: string;              
  attempted: boolean;        
  microShiftId: string;      
  alignmentFocus: string;    
  userId: string;            
  updatedAt: string;         
  tone: string;              
  id: string;                
  duration: number;          
  type:string; 
  prompt?:string;
  answer?:string
}